Split GamePanel to its own panel
