+++ source/bin_shell/private/CMakeLists.txt
@@ -3,6 +3,7 @@ target_sources(potato_shell PRIVATE
     "${CMAKE_CURRENT_LIST_DIR}/camera.h"
     "${CMAKE_CURRENT_LIST_DIR}/camera_controller.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/camera_controller.h"
+    "${CMAKE_CURRENT_LIST_DIR}/game_panel.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/main.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/scene.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/scene.h"
